+++ CHANGELOG.md
@@ -31,6 +31,7 @@ All notable changes to this project will be documented in this file.
 - Setting changes when value is empty on oxe-web-admin
 - Note update and modification button appearances on oxe-web-admin
 - Miscellanous UI on oxe-web-admin
+- public/get_public_article_content for html and markdown
 
 ## [1.14.7] - 2023-01-27